fix category and description in create collection api
1. make description optional in request payload
2. make category an enum in request payload

+++ server/src/collections/services.rs
@@ -14,7 +14,7 @@ pub async fn insert_new_collection(
         user_id,
         create_collections_request.name,
         create_collections_request.description,
-        create_collections_request.category,
+        i32::from(create_collections_request.category),
     )
     .fetch_one(pool)
     .await?;
